was working on getting the program working using Pouyas keypair (which we will use as a local keypair)

+++ solbond/programs/solbond/src/state/invariant_pool_meta.rs
@@ -21,7 +21,7 @@ pub struct InvariantPoolList {
 pub struct InvariantPoolAccount {
 
     // Position of pool in the global list of pools
-    pub idx: u32,
+    pub idx: u8,
     pub max_idx: u32,
 
     // Things we need for SWAP
